Fix critical thread-safety and overflow bugs, add regression tests
- Subscriber ctx race: move redisFree from thread to stop_subscriber
  after join, eliminating use-after-free/double-free on _sub.ctx
- Subscriber null deref: add null checks for reply elements before
  dereferencing str pointers in pub/sub message handler
- stop_subscriber guard: check thread.joinable() instead of _sub.run
  to handle threads that exited on their own (e.g. network error)
- RedisCache signed/unsigned: change firstIndexToCopy from int to
  size_t, preventing negative values from bypassing bounds check
- RAL_Time overflow: guard ms * NANOS_PER_MILLI against int64
  overflow for timestamps beyond ~year 2262

+++ src/RedisCache.hpp
@@ -53,7 +53,7 @@ public:
     return _lastWrite;
   }
 
-  RAL_Time copyReadBuffer(Type& destValue, int firstIndexToCopy = 0, int* pElementsCopied = nullptr)
+  RAL_Time copyReadBuffer(Type& destValue, size_t firstIndexToCopy = 0, int* pElementsCopied = nullptr)
   {
     std::shared_lock<std::shared_mutex> lk(_swapMutex);
 
@@ -70,7 +70,7 @@ public:
     }
 
     auto& src = _buffers.at(_readIndex);
-    if (firstIndexToCopy >= static_cast<int>(src.size()))
+    if (firstIndexToCopy >= src.size())
     {
       if (pElementsCopied) *pElementsCopied = 0;
       return RAL_Time();
